requirements.txt
asgiref==3.2.3
Django==3.0.1
djangorestframework==3.11.0
Pillow==6.2.1
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0